Tried ordered map to get deterministic outputs.
Failed. Not an use case, coz, even if I maintain everything with order. I don't want to use multiple goroutines deterministically, If I do so, my program would become synchronous.

+++ internal/dto/top_words_response.go
@@ -2,7 +2,7 @@ package dto
 
 type TopWordsResponseDto struct {
 	Word      string `json:"word"`
-	Frequency int64  `json:"frequency"`
+	Frequency int    `json:"frequency"`
 }
 
 type ResponseDto struct {
